handle remove badge location type filter
